Optimize fetching index information on System/Indices page
Instead of fetching index information multiple times per index, the new implementation
fetches the information in bulk and pulls the details out of that larger data structure.

This should fix the timeouts of the System/Indices page on setups with a large number
of indices and/or a slow Elasticsearch cluster.